requirements.txt
APScheduler==3.7.0
yagmail==0.14.247
requests==2.22.0
django_apscheduler==0.5.2
django_debug_toolbar==3.2
pandas==0.25.3
PyMySQL==0.9.3
redis==3.4.1
PyHDFS==0.3.1
snownlp==0.12.3
Django==3.0.3
phone==0.4.3
botocore==1.14.6
pyspark==3.1.1
beautifulsoup4==4.9.0
Pillow==7.0.0
django-simpleui==2021.4.3
djangorestframework==3.11.0
numpy==1.18.1
pandas==0.25.3
